drivers: serial: numicro_uart: remove in favor of numaker driver
The numaker uart driver is more fleshed out than the numicro uart driver so I'm replacing the latter with the former. I'm not convinced the numicro driver has actually ever been seriously used as the init function is hardcoded to only initialize the clock for the UART0 module. The only in-tree SOC that uses the numicro uart driver is the m48x.dtsi which I have updated and tested with a numaker-pfm-m487 board.

Signed-off-by: Scott Laboe <scottyl2.718@gmail.com>