drivers: usb: numaker: conditional epintsts use
Only use EPINTSTS if you have to. That is, if there are more than 12 endpoints there must be an EPINTSTS register because INTSTS only has enough bits to hold the interrupts state for 12 endpoints. The usbd peripheral on m48x microcontrollers doesn't have the EPINTSTS register because it only has 12 endpoints available so it uses 12 bits in the INTSTS to indicate whether data is available. The purpose of this commit is to allow the numaker_usbd driver to work with the m48x soc.

Signed-off-by: Scott Laboe <scottyl2.718@gmail.com>