drivers: gpio: ti: sn74hc595 remove reset requirement
Remove the requirement to have a reset pin for the SN74HC595 shift register. The reset pin doesn't have to be connected to the microcontroller for the shift register to operate. One reason for not wanting to add a line for the reset pin is if you are operating the shift register across an isolation barrier where it is much more expensive to do so.

Signed-off-by: Scott Laboe <scottyl2.718@gmail.com>